requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
rembg>=2.0.66
pillow>=10.0.0
numpy>=1.26.0
python-multipart>=0.0.9
pydantic-settings>=2.2.1
python-dotenv>=1.0.1        # Optional, for .env support

# Optional testing dependencies
pytest>=8.1.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
